https-upgrades/tentative/service-worker/service-worker.https.sub.html

<!DOCTYPE html>
<html>
  <head>
    <title>HTTPS Upgrades: Service Worker.</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/common/get-host-info.sub.js"></script>
    <script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>

  </head>
  <body>
    <script>
      promise_test(async t => {
        // This test opens a new tab and registers a service worker. The service worker navigates
        // the opened tab to an http URL. When Https-Upgrades is enabled, the navigated tab should
        // load over https. The tab notifies the service worker which in turn notifies this page.

        // Open an empty tab. This will be navigated to resources/popup.html by the service worker.
        // SW could also navigate this page away, but then we can't check for expected outcome
        // below.
        window.open("resources/empty.html");

        const scope = '/https-upgrades/tentative/';
        const registration = await service_worker_unregister_and_register(
            t, '/https-upgrades/tentative/service-worker.js', scope);
        t.add_cleanup(() => registration.unregister());
        await wait_for_state(t, registration.installing || registration.active, 'activated');

        navigator.serviceWorker.ready.then( registration => {
          registration.active.postMessage("start");
        });

        function promiseResolver(resolve) {
          navigator.serviceWorker.addEventListener('message', function onMessage(e) {
            var message = e.data.message;
            if (message === 'done') {
              // If Https-Upgrades worked, the worker opened an https:// URL with the same origin,
              // so we should get a valid windowclient instance.
              assert_true(e.data.windowClientIsValid);
              service_worker_unregister_and_done(t, scope);
              resolve(message);
            }
          });
        }

        var p = await new Promise(promiseResolver);
        assert_equals(p, 'done');
      });
    </script>
  </body>
</html>
